feat(client): Refactor and clean up state widget and make use of the new state definitions
